service/pyproject.toml
[project]
name = "sora-approval-service"
version = "0.1.0"
description = "Agentic SORA approval generator (Applicant mode, hackathon cut) on Microsoft Agent Framework"
requires-python = ">=3.11"
dependencies = [
    "agent-framework>=1.0.0b2",
    "agent-framework-devui>=1.0.0b2",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sse-starlette>=2.1.0",
    "pydantic>=2.8.0",
    "python-multipart>=0.0.9",
    "pypdf>=5.0.0",
    "python-docx>=1.1.2",
    "chromadb>=0.5.0",
    "pyyaml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["workflow", "agents", "tools", "knowledge"]
